🔨 Refactor: join, leave 했을 때 서버에게 알려주는 함수 hook으로 분리
